% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducing.R
\name{reproduce_figure}
\alias{reproduce_figure}
\title{Reproduce an individual figure}
\usage{
reproduce_figure(figure_number = c("1", "4", "5", "6", "7", "8", "9", "10",
  "11", "13", "14", "15", "16", "17", "18", "19", "20", "21", "23", "24", "A1"))
}
\arguments{
\item{figure_number}{character}
}
\description{
Reproduce an individual figure
}
