% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{figure1}
\alias{figure1}
\title{Saves last figure as png and as pdf}
\usage{
figure1(title_, p, sub_title = FALSE, title = !sub_title)
}
\arguments{
\item{title_}{filename for save optional title for the plot}

\item{p}{unevaluated expression that performs the plot}

\item{sub_title}{\code{title_} will be added as sub title to the plot if set}

\item{title}{\code{title_} will be added as title to the plot if set}
}
\value{
\code{p} evaluated
}
\description{
Saves last figure as png and as pdf
}
